refactor: File 전용 핸들러 분리
- common과 file 간 순환 참조 해결
