+++ src/main/java/com/gonghak98/v2/common/exception/GlobalExceptionHandler.java
@@ -1,13 +1,11 @@
 package com.gonghak98.v2.common.exception;
 
-import com.gonghak98.v2.file.exception.ExcelFileExceptionType;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.TypeMismatchException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.multipart.MaxUploadSizeExceededException;
 
 @Slf4j
 @RestControllerAdvice
@@ -16,15 +14,6 @@ public class GlobalExceptionHandler {
     @ExceptionHandler(BaseException.class)
     public ResponseEntity<ExceptionResponse> handleBaseException(BaseException exception) {
         final BaseExceptionType exceptionType = exception.exceptionType();
-        return ResponseEntity.status(exceptionType.httpStatus())
-                             .body(new ExceptionResponse(exceptionType.httpStatus().value(),
-                                                         exceptionType.errorCode(),
-                                                         exceptionType.errorMessage()));
-    }
-
-    @ExceptionHandler(MaxUploadSizeExceededException.class)
-    public ResponseEntity<ExceptionResponse> handleMaxUploadSizeExceededException(MaxUploadSizeExceededException exception) {
-        final BaseExceptionType exceptionType = ExcelFileExceptionType.EXCEED_EXCEL_FILE_SIZE;
         return ResponseEntity.status(exceptionType.httpStatus())
                              .body(new ExceptionResponse(exceptionType.httpStatus().value(),
                                                          exceptionType.errorCode(),
